pyproject.toml
[tool.poetry]
name = "django deep project"
version = "0.1.0"
description = ""
authors = ["kimpearl3599 <kimpearl3599@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.0"
django-ninja = "0.16.1"
black = "21.12b0"
django-stubs = "1.9.0"
mypy = "0.931"
httprequests = "^0.3"
isort = "5.10.1"
PyMySQL = "1.0.2"
types-PyMySQL = "1.0.6"
gunicorn = "20.1.0"
uvicorn = "0.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120


[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.9
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "sparta.settings"
